+++ src/modules/products/products.service.ts
@@ -199,7 +199,7 @@ export class ProductsService {
     };
   }
 
-  async setInactive(clearNotifications?: boolean): Promise<ResponseDto> {
+  async setInactive(clearNotifications?: boolean, skipSync?: boolean): Promise<ResponseDto> {
     // get all products with stock 0 and active true
     // Si clearNotifications es true, eliminar todas las notificaciones
     if (clearNotifications === true) {
@@ -207,8 +207,12 @@ export class ProductsService {
     } else {
       await this.notificationsService.deleteReadedNotifications();
     }
-    await this.createNotificationNoPublicado();
-    await this.mercadoLibreService.listProducts();
+
+    if (!skipSync) {
+      await this.createNotificationNoPublicado();
+      await this.mercadoLibreService.listProducts();
+    }
+
     // get all products with stock 0 and active true, but not deprecated
     const products = await this.productsRepository.find({
       where: { stock: 0, activo: true, deprecado: false },
@@ -241,7 +245,7 @@ export class ProductsService {
       );
       await this.googleLoggingService.log(
         'Productos procesados como inactivos',
-        { count: products.length, clearNotifications },
+        { count: products.length, clearNotifications, skipSync },
         'INFO',
         'setInactive',
         'products',